+++ HISTORY.rst
@@ -6,6 +6,8 @@
   memory allocation, file system buffers, and CPU cache lines.
 - Add the management command `dependencies_uninstall` to uninstall JavaScript
   and Google Font dependencies.
+- Move `mayan/settings/literals.py` to `mayan/literals.py` to prevent
+  circular dependencies.
 
 4.10.3 (2025-12-24)
 ===================